+++ app/history.odin
@@ -4,11 +4,11 @@ import visual "../visual"
 
 // History entry for navigation
 History_Entry :: struct {
-	center_x:           f64,
-	center_y:           f64,
-	zoom:               f64,
-	max_iterations:     u32,
-	palette:            string,
+	center_x:            f64,
+	center_y:            f64,
+	zoom:                f64,
+	max_iterations:      u64,
+	palette:             string,
 	use_smooth_coloring: bool,
 }
 
@@ -26,11 +26,11 @@ history_save :: proc(state: ^App_State) {
 
 	// Create new history entry
 	entry := History_Entry {
-		center_x           = state.center_x,
-		center_y           = state.center_y,
-		zoom               = state.zoom,
-		max_iterations     = state.max_iterations,
-		palette            = state.palette,
+		center_x            = state.center_x,
+		center_y            = state.center_y,
+		zoom                = state.zoom,
+		max_iterations      = state.max_iterations,
+		palette             = state.palette,
 		use_smooth_coloring = state.use_smooth_coloring,
 	}
